requirements.txt
asgiref==3.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-extensions==3.1.5
django-filter==21.1
django-hitcount==1.3.5
django-polymorphic==3.1.0
djangorestframework==3.13.1
fontawesomefree==6.1.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.3.7
oauthlib==3.2.0
Pillow==9.1.0
psycopg2==2.9.7
pycparser==2.21
pydot==1.4.2
//PyJW==2.0.8
